refactor(fetch): using fetch explicitly instead of services (strict requirement in the exercise)
